+++ cdl/tests/data.py
@@ -26,10 +26,10 @@ def get_test_signal(filename: str) -> cdl.obj.SignalObj:
     """Return test signal
 
     Args:
-        filename (str): Filename
+        filename: Filename
 
     Returns:
-        SignalObj: Signal object
+        Signal object
     """
     return cdl.obj.read_signal(get_test_fnames(filename)[0])
 
@@ -38,10 +38,10 @@ def get_test_image(filename: str) -> cdl.obj.ImageObj:
     """Return test image
 
     Args:
-        filename (str): Filename
+        filename: Filename
 
     Returns:
-        ImageObj: Image object
+        Image object
     """
     return cdl.obj.read_image(get_test_fnames(filename)[0])
 
@@ -52,11 +52,11 @@ def create_paracetamol_signal(
     """Create test signal (Paracetamol molecule spectrum)
 
     Args:
-        size (int | None): Size of the data. Defaults to None.
-        title (str | None): Title of the signal. Defaults to None.
+        size: Size of the data. Defaults to None.
+        title: Title of the signal. Defaults to None.
 
     Returns:
-        SignalObj: Signal object
+        Signal object
     """
     obj = cdl.obj.read_signal(get_test_fnames("paracetamol.txt")[0])
     if title is not None:
@@ -101,8 +101,8 @@ def add_gaussian_noise_to_signal(
     """Add Gaussian (Normal-law) random noise to data
 
     Args:
-        signal (cdl.obj.SignalObj): Signal object
-        p (GaussianNoiseParam | None): Gaussian noise parameters.
+        signal: Signal object
+        p: Gaussian noise parameters.
     """
     if p is None:
         p = GaussianNoiseParam()
@@ -121,20 +121,19 @@ def create_noisy_signal(
     """Create curve data, optionally noised
 
     Args:
-        noiseparam (GaussianNoiseParam | None): Noise parameters. Default:
-            None: No noise
-        newparam (cdl.obj.NewSignalParam | None): New signal parameters.
-            Default: Gaussian, size=500, xmin=-10, xmax=10
-        addparam (cdl.obj.GaussLorentzVoigtParam | None): Additional parameters.
-            Default: a=1.0, sigma=1.0, mu=0.0, ymin=0.0
-        title (str | None): Title of the signal. Default: None
-            If not None, overrides the title in newparam
-        noised (bool | None): If True, add noise to the signal.
-            Default: None (use noiseparam)
-            If True, eventually creates a new noiseparam if None
+        noiseparam: Noise parameters. Default: None: No noise
+        newparam: New signal parameters.
+         Default: Gaussian, size=500, xmin=-10, xmax=10
+        addparam: Additional parameters.
+         Default: a=1.0, sigma=1.0, mu=0.0, ymin=0.0
+        title: Title of the signal. Default: None
+         If not None, overrides the title in newparam
+        noised: If True, add noise to the signal.
+         Default: None (use noiseparam)
+         If True, eventually creates a new noiseparam if None
 
     Returns:
-        cdl.obj.SignalObj: Signal object
+        cdl.obj.Signal object
     """
     if newparam is None:
         newparam = cdl.obj.NewSignalParam()
@@ -157,12 +156,12 @@ def create_2d_steps_data(size: int, width: int, dtype: np.dtype) -> np.ndarray:
     """Creating 2D steps data for testing purpose
 
     Args:
-        size (int): Size of the data
-        width (int): Width of the steps
-        dtype (np.dtype): Data type
+        size: Size of the data
+        width: Width of the steps
+        dtype: Data type
 
     Returns:
-        np.ndarray: 2D data
+        2D data
     """
     data = np.zeros((size, size), dtype=dtype)
     value = 1
@@ -184,13 +183,13 @@ def create_2d_random(
     """Creating 2D Uniform-law random image
 
     Args:
-        size (int): Size of the data
-        dtype (np.dtype): Data type
-        level (float | None): Level of the random noise. Defaults to 0.1.
-        seed (int | None): Seed for random number generator. Defaults to 1.
+        size: Size of the data
+        dtype: Data type
+        level: Level of the random noise. Defaults to 0.1.
+        seed: Seed for random number generator. Defaults to 1.
 
     Returns:
-        np.ndarray: 2D data
+        2D data
     """
     rng = np.random.default_rng(seed)
     amp = np.iinfo(dtype).max * level
@@ -209,16 +208,16 @@ def create_2d_gaussian(
     """Creating 2D Gaussian (-10 <= x <= 10 and -10 <= y <= 10)
 
     Args:
-        size (int): Size of the data
-        dtype (np.dtype): Data type
-        x0 (float | None): x0. Defaults to 0.
-        y0 (float | None): y0. Defaults to 0.
-        mu (float | None): mu. Defaults to 0.0.
-        sigma (float | None): sigma. Defaults to 2.0.
-        amp (float | None): Amplitude. Defaults to None.
+        size: Size of the data
+        dtype: Data type
+        x0: x0. Defaults to 0.
+        y0: y0. Defaults to 0.
+        mu: mu. Defaults to 0.0.
+        sigma: sigma. Defaults to 2.0.
+        amp: Amplitude. Defaults to None.
 
     Returns:
-        np.ndarray: 2D data
+        2D data
     """
     xydata = np.linspace(-10, 10, size)
     x, y = np.meshgrid(xydata, xydata)
@@ -242,7 +241,7 @@ def get_laser_spot_data() -> list[np.ndarray]:
     testing laser spot image processing features
 
     Returns:
-        list[np.ndarray]: List of NumPy arrays
+        List of NumPy arrays
     """
     znoise = create_2d_random(2000, np.uint16)
     zgauss = create_2d_gaussian(2000, np.uint16, x0=2.0, y0=-3.0)
@@ -282,13 +281,13 @@ def get_peak2d_data(
     testing 2D peak detection or similar image processing features
 
     Args:
-        p (PeakDataParam | None): Peak data test image parameters. Defaults to None.
-        seed (int | None): Seed for random number generator. Defaults to None.
-        multi (bool | None): If True, multiple peaks are generated.
-            Defaults to False.
+        p: Peak data test image parameters. Defaults to None.
+        seed: Seed for random number generator. Defaults to None.
+        multi: If True, multiple peaks are generated.
+         Defaults to False.
 
     Returns:
-        np.ndarray: 2D data
+        2D data
     """
     if p is None:
         p = PeakDataParam()
@@ -320,11 +319,10 @@ def __set_default_size_dtype(
     """Set default shape and dtype
 
     Args:
-        p (cdl.obj.NewImageParam | None): Image parameters. Defaults to None.
-            If None, a new object is created.
+        p: Image parameters. Defaults to None. If None, a new object is created.
 
     Returns:
-        cdl.obj.NewImageParam: Image parameters
+        Image parameters
     """
     if p is None:
         p = cdl.obj.NewImageParam()
@@ -340,8 +338,8 @@ def add_gaussian_noise_to_image(
     """Add Gaussian noise to image
 
     Args:
-        src (cdl.obj.ImageObj): Source image
-        param (cdl.obj.NormalRandomParam): Parameters for the normal distribution
+        src: Source image
+        param: Parameters for the normal distribution
     """
     newparam = cdl.obj.new_image_param(
         height=image.data.shape[0],
@@ -359,10 +357,10 @@ def create_2dstep_image(
     """Creating 2D step image
 
     Args:
-        p (cdl.obj.NewImageParam | None): Image parameters. Defaults to None.
+        p: Image parameters. Defaults to None.
 
     Returns:
-        cdl.obj.ImageObj: Image object
+        Image object
     """
     p = __set_default_size_dtype(p)
     p.title = "Test image (2D step)" if p.title is None else p.title
@@ -388,15 +386,15 @@ def create_ring_data(
     """Create 2D ring data
 
     Args:
-        size (int): Size of the image
-        x0 (int): Center x coordinate
-        y0 (int): Center y coordinate
-        width (int): Width of the ring
-        radius (int): Radius of the ring
-        intensity (int): Intensity of the ring
+        size: Size of the image
+        x0: Center x coordinate
+        y0: Center y coordinate
+        width: Width of the ring
+        radius: Radius of the ring
+        intensity: Intensity of the ring
 
     Returns:
-        np.ndarray: 2D data
+        2D data
     """
     data = np.zeros((size, size), dtype=np.uint16)
     for x in range(data.shape[0]):
@@ -412,10 +410,10 @@ def create_ring_image(p: RingParam | None = None) -> cdl.obj.ImageObj:
     """Creating 2D ring image
 
     Args:
-        p (RingParam | None): Ring image parameters. Defaults to None.
+        p: Ring image parameters. Defaults to None.
 
     Returns:
-        cdl.obj.ImageObj: Image object
+        Image object
     """
     if p is None:
         p = RingParam()
@@ -440,10 +438,10 @@ def create_peak2d_image(
     """Creating 2D peak image
 
     Args:
-        p (cdl.obj.NewImageParam | None): Image parameters. Defaults to None.
+        p: Image parameters. Defaults to None.
 
     Returns:
-        cdl.obj.ImageObj: Image object
+        Image object
     """
     p = __set_default_size_dtype(p)
     p.title = "Test image (2D peaks)" if p.title is None else p.title
@@ -461,10 +459,10 @@ def create_sincos_image(
     """Creating test image (sin(x)+cos(y))
 
     Args:
-        p (cdl.obj.NewImageParam | None): Image parameters. Defaults to None.
+        p: Image parameters. Defaults to None.
 
     Returns:
-        cdl.obj.ImageObj: Image object
+        Image object
     """
     p = __set_default_size_dtype(p)
     p.title = "Test image (sin(x)+cos(y))" if p.title is None else p.title
@@ -480,24 +478,34 @@ def create_sincos_image(
 
 def create_noisygauss_image(
     p: cdl.obj.NewImageParam | None = None,
+    center: tuple[float, float] | None = None,
+    add_annotations: bool = True,
 ) -> cdl.obj.ImageObj:
     """Create test image (2D noisy gaussian)
 
     Args:
-        p (cdl.obj.NewImageParam | None): Image parameters. Defaults to None.
+        p: Image parameters. Defaults to None.
+        center: Center of the gaussian. Defaults to None.
+        add_annotations: If True, add annotations. Defaults to True.
 
     Returns:
-        ImageObj: Image object
+        Image object
     """
     p = __set_default_size_dtype(p)
     p.title = "Test image (noisy 2D Gaussian)" if p.title is None else p.title
     dtype = p.dtype.value
     size = p.width
     obj = cdl.obj.create_image_from_param(p)
-    obj.data = create_2d_gaussian(size, dtype=dtype, x0=2.0, y0=3.0) + create_2d_random(
+    if center is None:
+        # Default center
+        x0, y0 = 2.0, 3.0
+    else:
+        x0, y0 = center
+    obj.data = create_2d_gaussian(size, dtype=dtype, x0=x0, y0=y0) + create_2d_random(
         size, dtype
     )
-    obj.set_annotations_from_file(get_test_fnames("annotations.json")[0])
+    if add_annotations:
+        obj.set_annotations_from_file(get_test_fnames("annotations.json")[0])
     return obj
 
 
@@ -507,10 +515,10 @@ def create_multigauss_image(
     """Create test image (multiple 2D-gaussian peaks)
 
     Args:
-        p (cdl.obj.NewImageParam | None): Image parameters. Defaults to None.
+        p: Image parameters. Defaults to None.
 
     Returns:
-        ImageObj: Image object
+        Image object
     """
     p = __set_default_size_dtype(p)
     p.title = "Test image (multi-2D-gaussian)" if p.title is None else p.title
@@ -529,7 +537,7 @@ def create_annotated_image(title: str | None = None) -> cdl.obj.ImageObj:
     """Create test image with annotations
 
     Returns:
-        ImageObj: Image object
+        Image object
     """
     data = create_2d_gaussian(600, np.uint16, x0=2.0, y0=3.0)
     title = "Test image (with metadata)" if title is None else title
@@ -542,7 +550,7 @@ def create_resultshapes() -> tuple[cdl.obj.ResultShape, ...]:
     """Create test result shapes (core.model.base.ResultShape test objects)
 
     Returns:
-        tuple[ResultShape, ...]: Tuple of ResultShape objects
+        Tuple of ResultShape objects
     """
     RShape, SType = cdl.obj.ResultShape, cdl.obj.ShapeTypes
     return (